requirements.txt
gurobipy==10.0.1
matplotlib==3.7.0
networkx==3.1
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
scipy==1.9.1
seaborn==0.12.2
